% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_matrix.R
\name{as_population_matrix}
\alias{as_population_matrix}
\alias{as_population_matrix.social_contact_matrix}
\title{Coerce to a population contact matrix}
\usage{
as_population_matrix(x, ...)

\method{as_population_matrix}{social_contact_matrix}(x, population, ...)
}
\arguments{
\item{x}{An \R object..
Vector of grouping that represent the \emph{ith} (row) in \eqn{m_{ij}} contacts.}

\item{...}{Additional arguments to be passed to or from methods.}

\item{population}{\verb{[numeric]}.
Vector of length equal to the number of rows (columns) in \code{x}. The \emph{kth}
entry corresponds to the group represented by the \emph{kth} row of \code{x}.}
}
\description{
\code{as_population_matrix()} is a generic for coercing an input to a population
contact matrix.
}
