% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routine.R
\name{CJ.dt}
\alias{CJ.dt}
\title{CJ.dt CJ on data.table}
\usage{
CJ.dt(X, Y)
}
\arguments{
\item{X}{the first data table}

\item{Y}{the second data table}
}
\description{
CJ.dt CJ on data.table
}
\examples{
CJ.dt(data.table(X = 1:2), data.table(Y = c('a','b')))
}
