% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{returns a disambiguated match}
\usage{
clean(
  r,
  key_a,
  key_b,
  name_a,
  name_b,
  lf_a,
  lf_b,
  zip_a,
  zip_b,
  city_a,
  city_b,
  street_a,
  street_b,
  orb_reg,
  cc
)
}
\arguments{
\item{r}{matched dataset to be disambiguated}

\item{key_a}{indicates the name of the variable storing the identifier of the A dataset}

\item{key_b}{indicates the name of the variable storing the identifier of the B dataset}

\item{name_a}{indicates the name of the variable storing the original name before the normalization as stored in dataset A}

\item{name_b}{indicates the name of the variable storing the original name before the normalization as stored in dataset B}

\item{lf_a}{indicates the name of the variable storing the legal forms extracted from the name of entity stored in dataset A}

\item{lf_b}{indicates the name of the variable storing the legal forms extracted from the name of entity stored in dataset B}

\item{zip_a}{indicates the name of the variable storing the postal code of the entity stored in dataset A}

\item{zip_b}{indicates the name of the variable storing the postal code of the entity stored in dataset B}

\item{city_a}{indicates the name of the variable storing the name of the city of the entity stored in dataset A}

\item{city_b}{indicates the name of the variable storing the name of the city of the entity stored in dataset B}

\item{street_a}{indicates the name of the variable storing the street address of the entity stored in dataset A}

\item{street_b}{indicates the name of the variable storing the street address of the entity stored in dataset B}

\item{orb_reg}{indicates the name of the variable storing the information about the region of the entity, for instance NUTS3 in Orbis}

\item{cc}{indicates the name of the variable storing the information about the country code of the entity}
}
\value{
an original dataframe with additional variables with similarity scores
}
\description{
This function is being used after the match, to correct multiple matches. It uses additional features such as
original names, address and legal form information to assign similarity scores
}
