% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CM.R
\docType{package}
\name{CM}
\alias{CM}
\alias{CM-package}
\title{CM Package}
\description{
This package provides functionality to implement latent conjugate multivariate models for non-Gaussian data.
}
\author{
Jonathan R. Bradley \email{jrbradley@fsu.edu}
}
