% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_matrix.R
\name{contact_matrix}
\alias{contact_matrix}
\title{Contact matrix constructor}
\usage{
contact_matrix(
  from = list(),
  to = list(),
  value = numeric(),
  order = NULL,
  fill = 0
)
}
\arguments{
\item{from}{\verb{[list]}}

\item{to}{\verb{[list]}}

\item{value}{\verb{[numeric]}}

\item{order}{\verb{[character]}}

\item{fill}{\verb{[numeric]}}
}
\description{
\code{contact_matrix()} creates a social contract matrix from long form input.
}
