% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routine.R
\name{create_json_single}
\alias{create_json_single}
\title{Create json file for one part}
\usage{
create_json_single(idx, cnt, out_json_file, in_file)
}
\arguments{
\item{idx}{the rows to be put in this json file}

\item{cnt}{specifies the thread number}

\item{out_json_file}{names of the output json files}
}
\description{
Create json file for one part
}
