% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_address.R
\name{detect_address}
\alias{detect_address}
\title{returns a dataset enriched with additional variables indicating elements of address}
\usage{
detect_address(r, address, prefix, cc)
}
\arguments{
\item{r}{dataset containing freeline address data}

\item{address}{name of the column containing unstructured address data}

\item{prefix}{a string that will be addedd to the names of the column with found address elements}

\item{cc}{indicates a code of a country where the cities A and B are located

along with the original version of the name}
}
\value{
an original dataframe with ordered and ranked matched dataframe
}
\description{
This functions detects different elements of address in the unstructured address field and puts it into separate columns
}
