% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routine.R
\name{find_map_infty_step}
\alias{find_map_infty_step}
\title{find the final retired id and active id after infinity step}
\usage{
find_map_infty_step(map_one_step)
}
\description{
find the final retired id and active id after infinity step
}
\examples{
map_one_step <- data.table(retired_id = c('A','B','C', 'E'), active_id = c('B','C','D','F'))
map_infty_step <- find_map_infty_step(map_one_step)
}
