% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{\%+\%}
\alias{\%+\%}
\title{Concatenate two strings}
\usage{
a \%+\% b
}
\arguments{
\item{a:}{string in front}

\item{b:}{string in back}
}
\value{
concatenate string
}
\description{
Concatenate two strings
}
