% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_cities.R
\name{harmonize_cities}
\alias{harmonize_cities}
\title{returns a dataset with harmonized names of the cities}
\usage{
harmonize_cities(r, city_a, city_b, cc, new_col = T)
}
\arguments{
\item{r}{dataset containing names of the cities}

\item{city_a}{indicates a name of the column storing a name of city A}

\item{city_b}{indicates a name of the column storing a name of city B}

\item{cc}{indicates a code of a country where the cities A and B are located}

\item{new_col}{if TRUE creates new columns with normalized version of the cities A and B names
along with the original version of the name}
}
\value{
an original dataframe with ordered and ranked matched dataframe
}
\description{
This function allows for proper comparison of the cities' names if they come from different sources
}
