% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logitbetasim.R
\name{logitbetasim}
\alias{logitbetasim}
\title{Simulate logitBeta}
\usage{
logitbetasim(alpha, kappa)
}
\arguments{
\item{alpha}{A n-dimensional vector of shape parameters}

\item{kappa}{A second n-dimensional vector of shape parameters. kappa must be greater than alpha.}
}
\value{
W An n-dimensional vector of simulated values
}
\description{
This code simulates logit beta random variables
}
