% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MoransIBasis.R
\name{MoransI.Basis}
\alias{MoransI.Basis}
\title{Moran's I Basis Function}
\usage{
MoransI.Basis(X, r, A)
}
\arguments{
\item{X}{An nxp matrix of covariates. Each column represents a covariate. Each row corresponds to one of the n replicates.}

\item{r}{The number of basis functions.}

\item{A}{An nxn adjacency matrix.}
}
\value{
Psi nxr matrix of basis functions
}
\description{
This code computes the Moran's I basis functions
}
