% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmatch.R
\name{nmatch}
\alias{nmatch}
\title{Match records from different sources based on normalized names}
\usage{
nmatch(a, b, aname, bname, sort = T, unique = T)
}
\arguments{
\item{a}{first dataset to be matched}

\item{b}{second dataset to be matched}

\item{aname}{harmonized name in the first dataset}

\item{bname}{harmonized name in the second dataset}

\item{sort}{whether the words in harmonized names should be sorted alphabetically before matching}

\item{unique}{whether the duplicate names in harmonized names should be eliminated before matching}
}
\value{
a new dataframe only with matched pairs of names
}
\description{
This function matches records from different sources. It implements exact match method
}
\examples{
nmatch(x,y,"xname","yname",sort=T)
}
