% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norma.R
\name{norma}
\alias{norma}
\title{returns an original dataset with normalized name}
\usage{
norma(r, name, country, key, new_col = T, short = T, translit = T, legal = T)
}
\arguments{
\item{r}{dataset to be transformed}

\item{name}{name of the column storing name to be transformed, it has to be a string}

\item{country}{name of the column storing name of the country of seat}

\item{key}{name of the column storing unique identifier of the entity}

\item{new_col}{if TRUE creates new column with normalized version of the name along with the original version of the name}

\item{short}{if TRUE looks for 'trading as' expressions with the name}

\item{translit}{if TRUE looks for names written in cyrylic and transforms them into their latin versions}

\item{legal}{if TRUE looks for legal form expressions in the name and deals with them}
}
\value{
an original dataframe with two additional copies of records containing 'trading as' expression
}
\description{
This is the main function of the package
}
\examples{
df<-data.frame(name="Andrzej Beata Celina spółka z ograniczoną odpowiedzialnoscią nazwa skrócona ABC sp. z o.o. SP.J",
 country="PL",key=1)
}
