% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norma_legal.R
\name{norma_legal}
\alias{norma_legal}
\title{This function deals with legal forms indications present in the names}
\usage{
norma_legal(r, name, country)
}
\arguments{
\item{r}{dataset to be transformed}

\item{name}{name of the column storing name to be transformed, it has to be a string}

\item{country}{name of the column storing country of seat, it has to be a string}
}
\value{
an original dataframe with cleaned version of original name and additional column with standardized version of legal form
}
\description{
This function is developed to deal with the legal forms expressions found in the names
Treatment is country specific and the algorithm is not dealing with legal forms, which may
be associated with countries other than countries of seat.
}
\examples{
norma_legal(df,"name","country_code")
}
