% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norma_short.R
\name{norma_short}
\alias{norma_short}
\title{Identify and deal with 'trading as' expression in the name}
\usage{
norma_short(r, name)
}
\arguments{
\item{r}{dataset to be transformed}

\item{name}{name of the column storing name to be transformed, it has to be a string}
}
\value{
an original dataframe with two additional copies of records containing 'trading as' expression
}
\description{
This function is developed to deal with 'trading as' type of indications in the name. It identifies 'trading as' phrase
in different languages and creates three copies of record, for which this phrase has been identified: original version of name
a string preceding 'trading as' expression and a string following 'trading as' expression.
}
\examples{
norma_short(df,"name")
}
