% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norma_trans.R
\name{norma_trans}
\alias{norma_trans}
\title{Transliterate characters other than Latin}
\usage{
norma_trans(r, name, country, key)
}
\arguments{
\item{r}{dataset to be transformed}

\item{name}{name of the column storing name to be transformed, it has to be a string}

\item{country}{name of the column storing country of seat, it has to be a string}

\item{key}{name of the column storing key id, it has to be a string}
}
\value{
an original dataframe with transliterated version of name
}
\description{
This function is developed to transliterate the strings in other alphabets than Latin
}
\examples{
norma_trans(df,"name","country_code")
}
