% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_matrix.R
\name{population_contact_matrix}
\alias{population_contact_matrix}
\title{Create a population contact matrix}
\usage{
population_contact_matrix(
  i = character(),
  j = character(),
  value,
  groups = unique(c(i, j)),
  population,
  from = c("columns", "rows"),
  fill = 0
)
}
\arguments{
\item{i}{\verb{[character]}.
Vector of grouping that represent the \emph{ith} (row) in \eqn{C_{ij}} contacts.}

\item{j}{\verb{[character]}.
Vector of grouping that represent the \emph{jth} (column) in \eqn{C_{ij}} contacts.}

\item{value}{\verb{[numeric]}.
Vector of values corresponding corresponding to inputs \code{i} and \code{j}.}

\item{groups}{\verb{[character]}.
Possible groups that can appear in \code{i} and \code{j}. This will be used as the
\code{dimnames()} of the resulting contact matrix. By default this is set to
unique(c(i, j)).}

\item{population}{\verb{[numeric]}.
Vector of length the same as \code{groups} with value corresponding to the
population size of group with the \emph{kth} value corresponding to the \emph{kth}
entry in \code{groups}. Used to validate the correctness of the resulting
contact matrix, that is \eqn{C_{ij} * w_{j} == C_{ji} * w_{i}}, where
\eqn{w_{k}} represents the population in group \emph{k}.}

\item{from}{\verb{[character]}.
Should the contact matrix be interpreted as from "columns" to rows or from
"rows" to columns. Defaults to \code{from = "columns"}.}

\item{fill}{\verb{[numeric]}.
Numeric scalar used to fill empty values in the resulting contact matrix.
Defaults to 0.}
}
\description{
\code{population_contact_matrix()} creates a social contract matrix from \code{i}, \code{j}
and \code{value} vectors following the definition of given in \cite{Klepac et al}.
}
