% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank.R
\name{rankmatch}
\alias{rankmatch}
\title{returns a dataset with ordered and ranked matches}
\usage{
rankmatch(
  r,
  prefer_root = T,
  bvd_is_root = 1,
  name_sim = 1,
  lf_sim = 1,
  reg_sim = 1,
  city_sim = 1,
  street_sim = 1
)
}
\arguments{
\item{r}{cleaned dataset to be ranked}

\item{prefer_root}{logical variable indicating whether the ORBIS observations with the bvd_id number without hyphen should be preferred}

\item{bvd_is_root}{weight assigned to the criterion of bvd_id_number not to include hyphen}

\item{name_sim}{weight assigned to the similarity of original names}

\item{lf_sim}{weight assigned to the similarity of legal forms}

\item{reg_sim}{weight assigned to the similarity of seat regions as shown in respective datasets}

\item{city_sim}{weight assigned to the similarity of cities of seat as shown in respective datasets}

\item{street_sim}{weight assigned to the similarity of cities of seat as shown in respective datasets}
}
\value{
an original dataframe with ordered and ranked matched dataframe
}
\description{
This function is being used after the disambiguation of matches to rank possible candidates for the match
}
