% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routine.R
\name{sample.dt}
\alias{sample.dt}
\title{sample random rows from the data table X}
\usage{
sample.dt(X, nrow = 1, ...)
}
\arguments{
\item{X}{input data table}

\item{nrow}{number of rows to be sampled}
}
\description{
sample random rows from the data table X
}
\examples{
sample.dt(data.table(num = 1:10, lett = letters[1:10], Lett = LETTERS[1:10]))
}
