% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routine.R
\name{save_json}
\alias{save_json}
\title{Save dt into several json files}
\usage{
save_json(dt, out_folder = "out/", json_file_name = "file", nt = 5)
}
\arguments{
\item{dt}{the input data table}

\item{out_folder}{place to put output files}

\item{json_file_name}{names of the output json files}

\item{nt}{number of threads to use in parallel computing}
}
\description{
Save dt into several json files
}
\examples{
dt <- data.table(num = 1:26, lett = letters, LETT = LETTERS)
cm::save_json(dt, json_file_name = 'test', nt = 3)
jsonlite::read_json('out/file_2.json', simplifyVector = TRUE)
}
