% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routine.R
\name{split_into_n}
\alias{split_into_n}
\title{Split a vector x into a list of n}
\usage{
split_into_n(xx, n)
}
\description{
This function is pretty useful to use together with parallel computing,
e.g. when using clusterApply
}
\examples{
split_into_n(1:100,10)
split_into_n(letters[1:20],3)

}
