% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routine.R
\name{where}
\alias{where}
\title{Apply a boolean function to each entry of x}
\usage{
where(f, x)
}
\arguments{
\item{f}{a boolean function}

\item{x}{a numeric vector}
}
\value{
[(f(x_1), f(x_2), ..., f(x_n))]
}
\description{
[f(x_1), f(x_2), ..., f(x_n)]
}
\examples{
where(function(x) x>0, -2:2)
}
