% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTransform.R
\name{BTransform}
\alias{BTransform}
\title{Bayesian Transformations}
\usage{
BTransform(
  B,
  data_continuous,
  data_poisson,
  data_multinomial,
  nn,
  report = 100,
  nslice = 2
)
}
\arguments{
\item{B}{The number of iterations of the Gibbs sampler}

\item{data_continuous}{Normal distributed data vector}

\item{data_poisson}{Poisson distributed data vector}

\item{data_multinomial}{Binomial distributed data vector}

\item{nn}{The sample size assocated with the binomial observations. Must be same length of data_multinomial}

\item{report}{Option to print the iteration number of the MCMC.}

\item{nslice}{The burnin for the slice sampler}
}
\value{
ans A list of updated parameter values.
}
\description{
This code posterior replicates from the transformation model described in Bradley et al. (2020).
}
