% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_region.R
\name{detect_region}
\alias{detect_region}
\title{returns a dataset enriched with additional variables indicating region in the country}
\usage{
detect_region(zip, cc)
}
\arguments{
\item{zip}{name of the column containing zip code}

\item{cc}{indicates a code of a country where the cities A and B are located

along with the original version of the name}

\item{r}{a dataset containing a column with a zip code}
}
\value{
an original dataframe with a column describing region added
}
\description{
This function is using a zip code to return information on the NUTS3 region
}
