% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routine.R
\name{fastPOSIXctWrapper}
\alias{fastPOSIXctWrapper}
\title{An extention of fasttime::fastPOSIXct}
\usage{
fastPOSIXctWrapper(x, from_tz = NULL, to_tz = NULL,
  required.components = 3L)
}
\arguments{
\item{x}{strig vector to interpret as timestamps}

\item{from_tz}{the time zone x lies in}

\item{required.components}{same as fasttime::fastPOSIXct}

\item{the}{time zone we are converting to}
}
\description{
Adds flexibity to both input and output timezones
}
\examples{
fastPOSIXctWrapper("2017-04-29 20:26:11", from_tz = "America/New_York", to_tz = "UTC")
}
